Enable strict null check for overmind (#3423)
* Enable strict null check for overmind

* Updates

* Fix one more!

* typing fixes

* type fixes

* Convert live

* Convert client

* Add support for @babel/plugin-proposal-nullish-coalescing-operator in app

* Clean up yarn.lock

* Fix src/app/overmind/effects/git/export-to-github.ts

* Fix the loading of sandboxes

* more null check handling

* Fix type of color

* Fix typing of module id

* Change some typechecks on currentModule

* Fix now check

* Fix check for user.id

* Handle forking errors more gracefully

+++ packages/app/src/app/overmind/namespaces/editor/actions.ts
@@ -64,7 +64,7 @@ export const sandboxChanged: AsyncAction<{ id: string }> = withLoadApp<{
   id: string;
 }>(async ({ state, actions, effects }, { id }) => {
   // This happens when we fork. This can be avoided with state first routing
-  if (state.editor.isForkingSandbox) {
+  if (state.editor.isForkingSandbox && state.editor.currentSandbox) {
     effects.vscode.openModule(state.editor.currentModule);
 
     await actions.editor.internal.initializeLiveSandbox(
@@ -96,7 +96,6 @@ export const sandboxChanged: AsyncAction<{ id: string }> = withLoadApp<{
     const sandbox = await effects.api.getSandbox(newId);
 
     actions.internal.setCurrentSandbox(sandbox);
-
     actions.workspace.openDefaultItem();
   } catch (error) {
     state.editor.notFound = true;
@@ -109,7 +108,7 @@ export const sandboxChanged: AsyncAction<{ id: string }> = withLoadApp<{
     return;
   }
 
-  const sandbox = state.editor.currentSandbox;
+  const sandbox = state.editor.currentSandbox!;
 
   await effects.vscode.changeSandbox(sandbox, fs => {
     state.editor.modulesByPath = fs;
@@ -175,6 +174,10 @@ export const onOperationApplied: Action<{
   moduleShortid: string;
   code: string;
 }> = ({ state, effects, actions }, { code, moduleShortid }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   const module = state.editor.currentSandbox.modules.find(
     m => m.shortid === moduleShortid
   );
@@ -201,6 +204,10 @@ export const codeChanged: Action<{
   event?: any;
 }> = ({ effects, state, actions }, { code, event, moduleShortid }) => {
   effects.analytics.trackOnce('Change Code');
+
+  if (!state.editor.currentSandbox) {
+    return;
+  }
 
   const module = state.editor.currentSandbox.modules.find(
     m => m.shortid === moduleShortid
@@ -237,6 +244,10 @@ export const codeChanged: Action<{
 export const saveClicked: AsyncAction = withOwnedSandbox(
   async ({ state, effects, actions }) => {
     const sandbox = state.editor.currentSandbox;
+
+    if (!sandbox) {
+      return;
+    }
 
     try {
       const changedModules = sandbox.modules.filter(module =>
@@ -275,7 +286,7 @@ export const saveClicked: AsyncAction = withOwnedSandbox(
       });
 
       if (
-        state.editor.currentSandbox.originalGit &&
+        sandbox.originalGit &&
         state.workspace.openedWorkspaceItem === 'github'
       ) {
         actions.git.internal.fetchGitChanges();
@@ -292,6 +303,9 @@ export const saveClicked: AsyncAction = withOwnedSandbox(
 );
 
 export const createZipClicked: Action = ({ state, effects }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
   effects.zip.download(state.editor.currentSandbox);
 };
 
@@ -313,6 +327,10 @@ export const forkSandboxClicked: AsyncAction = async ({
   effects,
   actions,
 }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   if (
     state.editor.currentSandbox.owned &&
     !state.editor.currentSandbox.customTemplate &&
@@ -351,6 +369,10 @@ export const moduleSelected: Action<{
 
   const sandbox = state.editor.currentSandbox;
 
+  if (!sandbox) {
+    return;
+  }
+
   try {
     let module;
 
@@ -361,9 +383,7 @@ export const moduleSelected: Action<{
         sandbox.directories
       );
     } else {
-      module = state.editor.currentSandbox.modules.find(
-        moduleItem => moduleItem.id === id
-      );
+      module = sandbox.modules.find(moduleItem => moduleItem.id === id);
     }
 
     if (module.shortid === state.editor.currentModuleShortid) {
@@ -372,7 +392,7 @@ export const moduleSelected: Action<{
 
     actions.editor.internal.setCurrentModule(module);
 
-    if (state.live.isLive) {
+    if (state.live.isLive && state.live.liveUser && state.live.roomInfo) {
       effects.vscode.updateUserSelections(
         module,
         actions.live.internal.getSelectionsForModule(module)
@@ -449,6 +469,9 @@ export const prettifyClicked: AsyncAction = async ({
 }) => {
   effects.analytics.track('Prettify Code');
   const module = state.editor.currentModule;
+  if (!module.id) {
+    return;
+  }
   const newCode = await effects.prettyfier.prettify(
     module.id,
     module.title,
@@ -462,13 +485,18 @@ export const prettifyClicked: AsyncAction = async ({
 };
 
 export const errorsCleared: Action = ({ state, effects }) => {
+  const sandbox = state.editor.currentSandbox;
+  if (!sandbox) {
+    return;
+  }
+
   if (state.editor.errors.length) {
     state.editor.errors.forEach(error => {
       try {
         const module = resolveModule(
           error.path,
-          state.editor.currentSandbox.modules,
-          state.editor.currentSandbox.directories
+          sandbox.modules,
+          sandbox.directories
         );
         module.errors = [];
       } catch (e) {
@@ -493,6 +521,10 @@ export const frozenUpdated: AsyncAction<{ frozen: boolean }> = async (
   { state, effects },
   { frozen }
 ) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   state.editor.currentSandbox.isFrozen = frozen;
 
   await effects.api.saveFrozen(state.editor.currentSandbox.id, frozen);
@@ -525,6 +557,10 @@ export const discardModuleChanges: Action<{
   effects.analytics.track('Code Discarded');
 
   const sandbox = state.editor.currentSandbox;
+  if (!sandbox) {
+    return;
+  }
+
   const module = sandbox.modules.find(
     moduleItem => moduleItem.shortid === moduleShortid
   );
@@ -541,6 +577,10 @@ export const fetchEnvironmentVariables: AsyncAction = async ({
   state,
   effects,
 }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   state.editor.currentSandbox.environmentVariables = await effects.api.getEnvironmentVariables(
     state.editor.currentSandbox.id
   );
@@ -550,6 +590,10 @@ export const updateEnvironmentVariables: AsyncAction<EnvironmentVariable> = asyn
   { state, effects },
   environmentVariable
 ) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   state.editor.currentSandbox.environmentVariables = await effects.api.saveEnvironmentVariable(
     state.editor.currentSandbox.id,
     environmentVariable
@@ -561,6 +605,10 @@ export const updateEnvironmentVariables: AsyncAction<EnvironmentVariable> = asyn
 export const deleteEnvironmentVariable: AsyncAction<{
   name: string;
 }> = async ({ state, effects }, { name }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   const { id } = state.editor.currentSandbox;
 
   state.editor.currentSandbox.environmentVariables = await effects.api.deleteEnvironmentVariable(
@@ -594,6 +642,9 @@ export const previewActionReceived: Action<any> = (
       });
       break;
     case 'show-error': {
+      if (!state.editor.currentSandbox) {
+        return;
+      }
       const error: ModuleError = {
         column: action.column,
         line: action.line,
@@ -622,6 +673,9 @@ export const previewActionReceived: Action<any> = (
       break;
     }
     case 'show-correction': {
+      if (!state.editor.currentSandbox) {
+        return;
+      }
       const correction: ModuleCorrection = {
         path: action.path,
         column: action.column,
@@ -634,7 +688,7 @@ export const previewActionReceived: Action<any> = (
       };
       try {
         const module = resolveModule(
-          correction.path,
+          correction.path as string,
           state.editor.currentSandbox.modules,
           state.editor.currentSandbox.directories
         );
@@ -648,6 +702,10 @@ export const previewActionReceived: Action<any> = (
       break;
     }
     case 'clear-errors': {
+      const sandbox = state.editor.currentSandbox;
+      if (!sandbox) {
+        return;
+      }
       const currentErrors = state.editor.errors;
 
       const newErrors = clearCorrectionsFromAction(currentErrors, action);
@@ -657,8 +715,8 @@ export const previewActionReceived: Action<any> = (
           try {
             const module = resolveModule(
               error.path,
-              state.editor.currentSandbox.modules,
-              state.editor.currentSandbox.directories
+              sandbox.modules,
+              sandbox.directories
             );
 
             module.errors = [];
@@ -669,8 +727,8 @@ export const previewActionReceived: Action<any> = (
         newErrors.forEach(error => {
           const module = resolveModule(
             error.path,
-            state.editor.currentSandbox.modules,
-            state.editor.currentSandbox.directories
+            sandbox.modules,
+            sandbox.directories
           );
 
           module.errors.push(error);
@@ -681,6 +739,12 @@ export const previewActionReceived: Action<any> = (
       break;
     }
     case 'clear-corrections': {
+      const sandbox = state.editor.currentSandbox;
+
+      if (!sandbox) {
+        return;
+      }
+
       const currentCorrections = state.editor.corrections;
 
       const newCorrections = clearCorrectionsFromAction(
@@ -692,9 +756,9 @@ export const previewActionReceived: Action<any> = (
         state.editor.corrections.forEach(correction => {
           try {
             const module = resolveModule(
-              correction.path,
-              state.editor.currentSandbox.modules,
-              state.editor.currentSandbox.directories
+              correction.path!,
+              sandbox.modules,
+              sandbox.directories
             );
 
             module.corrections = [];
@@ -705,9 +769,9 @@ export const previewActionReceived: Action<any> = (
         });
         newCorrections.forEach(correction => {
           const module = resolveModule(
-            correction.path,
-            state.editor.currentSandbox.modules,
-            state.editor.currentSandbox.directories
+            correction.path!,
+            sandbox.modules,
+            sandbox.directories
           );
 
           module.corrections.push(correction);
@@ -719,6 +783,9 @@ export const previewActionReceived: Action<any> = (
     }
     case 'source.module.rename': {
       const sandbox = state.editor.currentSandbox;
+      if (!sandbox) {
+        return;
+      }
       const module = effects.utils.resolveModule(
         action.path.replace(/^\//, ''),
         sandbox.modules,
@@ -752,6 +819,9 @@ export const renameModule: AsyncAction<{
 }> = withOwnedSandbox(
   async ({ state, actions, effects }, { title, moduleShortid }) => {
     const sandbox = state.editor.currentSandbox;
+    if (!sandbox) {
+      return;
+    }
     const module = sandbox.modules.find(
       moduleItem => moduleItem.shortid === moduleShortid
     );
