+++ packages/app/src/app/overmind/namespaces/workspace/actions.ts
@@ -1,10 +1,10 @@
 import getTemplate from '@codesandbox/common/lib/templates';
 import { CustomTemplate } from '@codesandbox/common/lib/types';
+import track from '@codesandbox/common/lib/utils/analytics';
 import slugify from '@codesandbox/common/lib/utils/slugify';
 import { Action, AsyncAction } from 'app/overmind';
 import { withOwnedSandbox } from 'app/overmind/factories';
 import getItems from 'app/overmind/utils/items';
-import track from '@codesandbox/common/lib/utils/analytics';
 
 export const valueChanged: Action<{
   field: string;
@@ -21,6 +21,10 @@ export const tagAdded: AsyncAction = withOwnedSandbox(
   async ({ state, effects, actions }) => {
     const { tagName } = state.workspace.tags;
     const sandbox = state.editor.currentSandbox;
+
+    if (!sandbox) {
+      return;
+    }
 
     const cleanTag = tagName.replace(/#/g, '');
 
@@ -41,6 +45,11 @@ export const tagAdded: AsyncAction = withOwnedSandbox(
 export const tagRemoved: AsyncAction<string> = withOwnedSandbox(
   async ({ state, effects, actions }, tag) => {
     const sandbox = state.editor.currentSandbox;
+
+    if (!sandbox) {
+      return;
+    }
+
     const tagIndex = sandbox.tags.indexOf(tag);
 
     sandbox.tags.splice(tagIndex, 1);
@@ -48,8 +57,21 @@ export const tagRemoved: AsyncAction<string> = withOwnedSandbox(
     try {
       sandbox.tags = await effects.api.deleteTag(sandbox.id, tag);
 
+      if (
+        !state.editor.parsedConfigurations ||
+        !state.editor.currentPackageJSON
+      ) {
+        return;
+      }
       // Create a "joint action" on this
+      if (!state.editor.parsedConfigurations.package) {
+        return;
+      }
       const { parsed } = state.editor.parsedConfigurations.package;
+
+      if (!parsed) {
+        return;
+      }
 
       parsed.keywords = sandbox.tags;
       parsed.name = slugify(sandbox.title || sandbox.id);
@@ -74,23 +96,32 @@ export const tagsChanged: AsyncAction<{
   newTags: string[];
   removedTags: string[];
 }> = async ({ actions, effects, state }, { newTags, removedTags }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   const { tags } = state.editor.currentSandbox;
   if (tags.length > 5) {
-    return effects.notificationToast.error('You can have a maximum of 5 tags');
+    effects.notificationToast.error('You can have a maximum of 5 tags');
+    return;
   }
 
   const tagWasRemoved =
     newTags.length < tags.length && removedTags.length === 1;
   if (tagWasRemoved) {
-    return removedTags.forEach(actions.workspace.tagRemoved);
+    removedTags.forEach(actions.workspace.tagRemoved);
+    return;
   }
 
-  return actions.workspace.tagAdded();
+  await actions.workspace.tagAdded();
 };
 
 export const sandboxInfoUpdated: AsyncAction = withOwnedSandbox(
   async ({ state, effects, actions }) => {
     const sandbox = state.editor.currentSandbox;
+    if (!sandbox) {
+      return;
+    }
     const { project } = state.workspace;
     const hasChangedTitle =
       project.title.trim() && sandbox.title !== project.title;
@@ -122,6 +153,11 @@ export const sandboxInfoUpdated: AsyncAction = withOwnedSandbox(
           alias: project.alias,
         });
 
+        if (!updatedSandbox) {
+          effects.notificationToast.error('Could not update Sandbox');
+          return;
+        }
+
         effects.router.replaceSandboxUrl(updatedSandbox);
 
         await actions.editor.internal.updateSandboxPackageJson();
@@ -138,6 +174,9 @@ export const sandboxInfoUpdated: AsyncAction = withOwnedSandbox(
 
 export const externalResourceAdded: AsyncAction<string> = withOwnedSandbox(
   async ({ effects, state, actions }, resource) => {
+    if (!state.editor.currentSandbox) {
+      return;
+    }
     const { externalResources } = state.editor.currentSandbox;
 
     externalResources.push(resource);
@@ -166,6 +205,10 @@ export const externalResourceAdded: AsyncAction<string> = withOwnedSandbox(
 
 export const externalResourceRemoved: AsyncAction<string> = withOwnedSandbox(
   async ({ effects, state, actions }, resource) => {
+    if (!state.editor.currentSandbox) {
+      return;
+    }
+
     const { externalResources } = state.editor.currentSandbox;
     const resourceIndex = externalResources.indexOf(resource);
 
@@ -207,6 +250,10 @@ export const sandboxDeleted: AsyncAction = async ({
 }) => {
   actions.modalClosed();
 
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   await effects.api.deleteSandbox(state.editor.currentSandbox.id);
 
   // Not sure if this is in use?
@@ -220,6 +267,10 @@ export const sandboxPrivacyChanged: AsyncAction<{
   privacy: 0 | 1 | 2;
   source?: string;
 }> = async ({ actions, effects, state }, { privacy, source = 'generic' }) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   track('Sandbox - Update Privacy', {
     privacy,
     source,
@@ -278,6 +329,12 @@ export const deleteTemplate: AsyncAction = async ({
   effects,
 }) => {
   effects.analytics.track('Template - Removed', { source: 'editor' });
+  if (
+    !state.editor.currentSandbox ||
+    !state.editor.currentSandbox.customTemplate
+  ) {
+    return;
+  }
   const sandboxId = state.editor.currentSandbox.id;
   const templateId = state.editor.currentSandbox.customTemplate.id;
 
@@ -301,6 +358,10 @@ export const editTemplate: AsyncAction<CustomTemplate> = async (
   { state, actions, effects },
   template
 ) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   effects.analytics.track('Template - Edited', { source: 'editor' });
 
   const sandboxId = state.editor.currentSandbox.id;
@@ -327,6 +388,10 @@ export const addedTemplate: AsyncAction<{
   description: string;
   title: string;
 }> = async ({ state, actions, effects }, template) => {
+  if (!state.editor.currentSandbox) {
+    return;
+  }
+
   effects.analytics.track('Template - Created', { source: 'editor' });
 
   const sandboxId = state.editor.currentSandbox.id;
